Refactored util functions into new helper functions
